pyproject.toml
[project]
name = "canada-xc"
version = "0.1.0"
description = "Django application for Canadian cross country running"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django-htmx>=1.21.0",
    "django-import-export>=4.1.1",
    "django>=5.1.2",
    "gunicorn>=23.0.0",
    "pillow>=11.0.0",
    "psycopg[c]>=3.2.3",
    "python-dotenv>=1.0.1",
    "redis>=5.2.0",
    "sentry-sdk[django]>=2.18.0",
    "sorl-thumbnail>=12.11.0",
    "thefuzz>=0.22.1",
]

[tool.uv]
dev-dependencies = [
    "bs4>=0.0.2",
    "djade>=1.3.0",
    "django-debug-toolbar>=4.4.6",
    "ruff>=0.6.9",
]

[tool.ruff]
exclude = ["**/migrations/*.py", "settings.py"]

[tool.ruff.lint]
select = [
    "E",   # pycodestyle
    "F",   # Pyflakes
    "UP",  # pyupgrade
    "B",   # flake8-bugbear
    "SIM", # flake8-simplify
    "I",   # isort
]

# isort
[tool.ruff.lint.isort]
known-first-party = ["canadaxc"]
section-order = [
    "future",
    "standard-library",
    "third-party",
    "django",
    "first-party",
    "local-folder",
]

[tool.ruff.lint.isort.sections]
"django" = ["django"]
